fix: Make CollectionSnapshot extend AbstractList test: Add tests for CollectionSnapshot.get(index) chore: Remove redundant ListSnapshot, ConcurrentListSerializer and ConcurrentListSerializer
